redesign of adding new impulse
